Refactor intp pipeline folder
Move box, position and overlap files
