+++ index.js
@@ -1,3 +1,3 @@
 export { default } from './src/WeekView/WeekView';
 export { addLocale, createFixedWeekDate } from './src/utils/dates';
-export { OVERLAP_METHOD } from './src/Events/overlap';
+export { OVERLAP_METHOD } from './src/pipeline/overlap';
