+++ src/pipeline/overlap.js
